home-threat-detection/home-threat-detection/pyproject.toml
[project]
name = "home-threat-detection"
version = "0.1.0"
description = "Agentic Home Threat Detection System"
authors = [
    {name = "Your Name", email = "your@email.com"},
]
dependencies = [
    "google-adk[a2a]~=1.15.1",
    "google-cloud-aiplatform~=1.106.0",
    "fastapi~=0.115.8",
    "uvicorn~=0.34.0",
    "opencv-python~=4.10.0",
    "pillow~=10.4.0",
    "pinecone-client~=5.0.0",
    "numpy~=1.26.0",
    "pydantic~=2.9.0",
    "python-multipart~=0.0.9",
    "opentelemetry-exporter-gcp-trace~=1.9.0",
    "google-cloud-logging~=3.11.4",
]

requires-python = ">=3.10,<3.14"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.23.8",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"